segregated live build
