+++ pipeline/src/main/java/com/autobuild/pipeline/executor/controller/PipelineBuildController.java
@@ -1,7 +1,7 @@
 package com.autobuild.pipeline.executor.controller;
 
-import java.util.List;
 import java.io.IOException;
+import java.util.List;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 import com.autobuild.pipeline.executor.dto.PipelineBuildDTO;
 import com.autobuild.pipeline.executor.service.PipelineBuildService;
@@ -33,18 +32,6 @@ public class PipelineBuildController {
 
     @Autowired
     private PipelineBuildService service;
-
-    @GetMapping("/sse/subscribe/{pipelineBuildId}")
-    public SseEmitter getLivePipelineBuild(@PathVariable UUID pipelineBuildId) {
-        log.info("Subscription requested");
-        SseEmitter emitter = new SseEmitter(600000L);
-
-        service.addSubscriber(emitter, pipelineBuildId);
-
-        log.info("Subscription done. returning emitter");
-
-        return emitter;
-    }
 
     @GetMapping("/{pipelineBuildId}")
     public ResponseEntity<PipelineBuildDTO> getPipelineBuild(@PathVariable UUID pipelineBuildId) {
